Start refactoring subscription handling to use manager/worker model
